fix(test): restore locked domain entrypoints
Expose split routing contract and phrase cases through the Todo 1 module,
and compose lifecycle isolation cases into the Todo 10/F3 module. Keep the
split sources as non-TestCase mixins so full discovery executes every
scenario exactly once.

Constraint: Preserve exact Todo 1, Todo 5, Todo 8, Todo 10, and F3 unittest entrypoints with split test sources below 250 pure LOC.
Rejected: Re-export TestCase subclasses from split modules | full discovery would execute the same scenarios more than once.
Confidence: high
Scope-risk: narrow
Directive: Keep compatibility classes in locked entrypoint modules and split scenario containers outside TestCase inheritance.
Tested: exact Todo 1/5/8/10/F3 and dotted journey; 151 domain tests; 2816 unique discovery IDs; full 2816-test evaluator; compileall; docs check; Ruff; pure LOC; diff check.
Not-tested: Remote CI was not run.
