+++ tests/test_domain_context_lifecycle.py
@@ -1,211 +1,32 @@
 from __future__ import annotations
 
-from contextlib import contextmanager
-import json
-import os
 from pathlib import Path
-import socket
-import subprocess
-import sys
 from tempfile import TemporaryDirectory
 import unittest
-from unittest import mock
 
-from _local_package import load_local_package
-
-load_local_package()
-
-from test_domain_context_lifecycle_journey import DomainContextLifecycleJourneyMixin
-
-PUBLIC_CONTEXT_KEYS = {
-    "schema_version",
-    "workflow_hint",
-    "required_input",
-    "question",
-    "digest",
-    "claim_boundary",
-}
-REPOSITORY_ROOT = Path(__file__).resolve().parents[1]
-EN_SALES_QUESTION = (
-    "Which account or customer segment should this sales work focus on?"
+from domain_context_lifecycle_support import (
+    EN_SALES_QUESTION as EN_SALES_QUESTION,
+    PUBLIC_CONTEXT_KEYS as PUBLIC_CONTEXT_KEYS,
+    _approve_candidate as _approve_candidate,
+    _assert_applied_context as _assert_applied_context,
+    _block_external_connections as _block_external_connections,
+    _canonical,
+    _capture_and_approve,
+    _capture_candidate as _capture_candidate,
+    _chat,
+    _repository,
+    _run_omh as _run_omh,
+    _unresolved_message,
 )
-def _canonical(value: object) -> str:
-    return json.dumps(value, sort_keys=True, separators=(",", ":"), ensure_ascii=False)
+from test_domain_context_lifecycle_journey import DomainContextLifecycleJourneyMixin
+from test_domain_context_store_isolation import DomainContextStoreIsolationMixin
 
 
-def _repository(path: Path) -> Path:
-    path.mkdir(parents=True)
-    (path / ".git").mkdir()
-    return path.resolve()
-
-
-def _command_environment(root: Path) -> dict[str, str]:
-    environment = os.environ.copy()
-    environment["HOME"] = str(root.parent / "isolated-home")
-    environment["OMH_HOME"] = str(root.parent / "hostile-env" / ".omh")
-    environment["HERMES_HOME"] = str(root.parent / "hostile-env" / ".hermes")
-    environment["PYTHONPATH"] = os.pathsep.join(
-        (str(REPOSITORY_ROOT / "src"), str(REPOSITORY_ROOT / "tests"))
-    )
-    environment["NO_PROXY"] = "*"
-    environment["no_proxy"] = "*"
-    return environment
-
-
-def _run_omh(root: Path, *arguments: str) -> dict[str, object]:
-    result = subprocess.run(
-        [sys.executable, "-m", "omh.cli", *arguments],
-        cwd=root,
-        env=_command_environment(root),
-        text=True,
-        capture_output=True,
-        check=False,
-    )
-    if result.returncode != 0:
-        raise AssertionError(
-            f"omh command failed ({result.returncode}): {arguments!r}\n{result.stderr}"
-        )
-    try:
-        payload = json.loads(result.stdout)
-    except json.JSONDecodeError as exc:
-        raise AssertionError(
-            f"omh command did not return JSON: {arguments!r}\n{result.stdout}"
-        ) from exc
-    if not isinstance(payload, dict):
-        raise AssertionError(f"omh command returned a non-object: {arguments!r}")
-    return payload
-
-
-def _capture_candidate(
-    root: Path,
-    *,
-    scope_kind: str,
-    scope_ref: str,
-    domain: str,
-    phrase: str,
-    canonical: str,
-    workflow_hints: tuple[str, ...],
-) -> str:
-    arguments = [
-        "--scope",
-        "project",
-        "memory",
-        "domain-capture",
-        "--scope-kind",
-        scope_kind,
-        "--scope-ref",
-        scope_ref,
-        "--domain",
-        domain,
-        "--mapping",
-        f"{phrase}={canonical}",
-        "--source-ref",
-        "lifecycle-test",
-    ]
-    for hint in workflow_hints:
-        arguments.extend(("--workflow-hint", hint))
-    payload = _run_omh(root, *arguments)
-    candidate_id = str(payload["candidate"]["candidate_id"])
-    if not candidate_id:
-        raise AssertionError("public capture returned an empty candidate id")
-    return candidate_id
-
-
-def _approve_candidate(root: Path, candidate_id: str) -> dict[str, object]:
-    review = _run_omh(
-        root,
-        "--scope",
-        "project",
-        "memory",
-        "domain-review",
-        "--candidate",
-        candidate_id,
-    )
-    if review["cards"][0]["candidate_id"] != candidate_id:
-        raise AssertionError("public review did not return the captured candidate")
-    return _run_omh(
-        root,
-        "--scope",
-        "project",
-        "memory",
-        "domain-approve",
-        candidate_id,
-        "--approved-by",
-        "lifecycle-test",
-    )
-
-
-def _capture_and_approve(
-    root: Path,
-    *,
-    scope_kind: str = "project",
-    scope_ref: str | None = None,
-    domain: str,
-    phrase: str,
-    canonical: str = "review_marker",
-    workflow_hints: tuple[str, ...] = ("sales-development",),
-) -> dict[str, object]:
-    candidate_id = _capture_candidate(
-        root,
-        scope_kind=scope_kind,
-        scope_ref=scope_ref or root.name,
-        domain=domain,
-        phrase=phrase,
-        canonical=canonical,
-        workflow_hints=workflow_hints,
-    )
-    return _approve_candidate(root, candidate_id)
-
-
-def _chat(root: Path, message: str) -> dict[str, object]:
-    return _run_omh(
-        root,
-        "--scope",
-        "project",
-        "chat",
-        "interact",
-        "--source",
-        "discord",
-        "--json",
-        message,
-    )
-
-
-def _unresolved_message(phrase: str) -> str:
-    return f"something {phrase} feels unresolved"
-
-
-def _assert_applied_context(
-    testcase: unittest.TestCase,
-    interaction: dict[str, object],
-    *,
-    workflow: str,
-    locale: str,
-    question: str,
-) -> None:
-    testcase.assertEqual(set(interaction["domain_routing_context"]), PUBLIC_CONTEXT_KEYS)
-    context = interaction["domain_routing_context"]
-    testcase.assertEqual(context["schema_version"], "domain_routing_context/v1")
-    testcase.assertEqual(context["workflow_hint"], workflow)
-    testcase.assertEqual(context["question"], {"locale": locale, "text": question})
-    testcase.assertRegex(context["digest"], r"^[0-9a-f]{64}$")
-    testcase.assertEqual(
-        interaction["chat_response"]["body"].splitlines()[-1],
-        question,
-    )
-
-
-@contextmanager
-def _block_external_connections():
-    with mock.patch.object(
-        socket.socket,
-        "connect",
-        side_effect=AssertionError("external connection attempted"),
-    ):
-        yield
-
-
-class DomainContextLifecycleTests(DomainContextLifecycleJourneyMixin, unittest.TestCase):
+class DomainContextLifecycleTests(
+    DomainContextLifecycleJourneyMixin,
+    DomainContextStoreIsolationMixin,
+    unittest.TestCase,
+):
     def test_unrelated_malformed_health_artifact_suppresses_context(self) -> None:
         with TemporaryDirectory() as temporary:
             root = _repository(Path(temporary) / "unhealthy-project")
